# api/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/stores/{id}/generate": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Generate store files",
                "description": "Generate static HTML/CSS files for the store based on template and customization",
                "operationId": "037b8a3fa77f2c04c5d989d21298f85a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Store ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Store generated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "store_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/stores/my-store"
                                                },
                                                "files_generated": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Store not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to generate store",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "type": "object",
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error description"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "description": "Validation errors if applicable"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token obtained from login endpoint"
            }
        }
    },
...
    "info": {
        "title": "LivePetal E-Commerce Platform API",
        "description": "Multi-tenant e-commerce platform with JWT authentication. This API allows super admins to manage clients and their stores, while clients can manage their products and orders.",
        "version": "2.0.0",
        "contact": {
            "name": "API Support",
            "email": "support@livepetal.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.livepetal.com",
            "description": "Production Server"
        }
    ]
}